requirements.txt
ffmpeg-python==0.2.0 
google-generativeai==0.8.5
loguru==0.7.3 
load_dotenv
asyncio
aiohttp